Extract parseValue to utils
