+++ solidity/random-beacon/tasks/send-eth.ts
@@ -1,8 +1,8 @@
 import { task, types } from "hardhat/config"
 
 import type { BigNumber, Signer } from "ethers"
-import type { HardhatRuntimeEnvironment } from "hardhat/types"
 import type { TransactionResponse } from "@ethersproject/abstract-provider"
+import { parseValue } from "./utils"
 
 // eslint-disable-next-line import/prefer-default-export
 export const TASK_SEND_ETH = "send-eth"
@@ -45,16 +45,3 @@ task(TASK_SEND_ETH, "Send ether to an address")
 
     await tx.wait()
   })
-
-export function parseValue(
-  value: string,
-  hre: HardhatRuntimeEnvironment
-): BigNumber {
-  const parsed = String(value).trim().split(" ")
-
-  if (parsed.length === 0 || parsed.length > 2) {
-    throw new Error(`invalid value: ${value}`)
-  }
-
-  return hre.ethers.utils.parseUnits(parsed[0], parsed[1] || "wei")
-}
