+++ solidity/random-beacon/tasks/utils/index.ts
@@ -1,4 +1,5 @@
 export * from "./authorize"
+export * from "./ether"
 export * from "./mint"
 export * from "./register"
 export * from "./stake"
